requirements.txt
# requirements.txt (versión mejorada)
Django==4.2.3
mysqlclient==2.1.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.0
coreapi==2.3.3
cloudinary==1.32.0
django-cloudinary-storage==0.3.0

# Dependencias recomendadas para producción (añade estas)
waitress==2.1.2  # Servidor ligero para producción
python-dotenv==1.0.0  # Para manejar variables de entorno